finalize invoice parser refactoring
